requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.1
pydantic==2.11.5
python-dotenv==1.1.0
pytest==7.4.3
httpx==0.25.2
